pyproject.toml
[tool.poetry]
name = "beddybai"
version = "0.1.0"
description = ""
authors = ["David Young <dmyoung9@asu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.8.4"
openai = "^0.27.4"
tiktoken = "^0.3.3"
quart = "^0.18.4"
python-dotenv = "^1.0.0"
quart-sqlalchemy = "^2.0.0"
pydantic = "^1.10.7"
langchain = "^0.0.209"
sqlalchemy-serializer = "^1.4.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.263"
black = "^23.3.0"
pre-commit = "^3.2.2"
mypy = "^1.2.0"
parameterized = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
